# backend/api-docs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ReelNet API",
    "description" : "RESTful API for the ReelNet application",
    "contact" : {
      "name" : "ReelNet Support",
      "url" : "https://www.reelnet.com",
      "email" : "support@reelnet.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Server URL"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
...
  "paths" : {
    "/v1/vocabulary-sets/{id}" : {
      "get" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Get vocabulary set by ID",
        "description" : "Retrieves a vocabulary set by its ID if the user has access to it",
        "operationId" : "getVocabularySetById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVocabularySetDto"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Update vocabulary set",
        "description" : "Updates an existing vocabulary set if the user has access to it",
        "operationId" : "updateVocabularySet",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVocabularySetDto"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Delete vocabulary set",
        "description" : "Deletes a vocabulary set if the user has access to it",
        "operationId" : "deleteVocabularySet",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/v1/vocabulary-sets" : {
      "get" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Get all vocabulary sets",
        "description" : "Retrieves all vocabulary sets with pagination, filtering and searching",
        "operationId" : "getAllVocabularySets",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Page number (0-based)",
          "schema" : {
            "type" : "integer",
            "default" : 0
          },
          "example" : 0
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Page size",
          "schema" : {
            "type" : "integer",
            "default" : 10
          },
          "example" : 10
        }, {
          "name" : "sortBy",
          "in" : "query",
          "description" : "Field to sort by",
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          },
          "example" : "name"
        }, {
          "name" : "sortDirection",
          "in" : "query",
          "description" : "Sort direction",
          "schema" : {
            "type" : "string",
            "default" : "DESC",
            "enum" : [ "ASC", "DESC" ]
          },
          "example" : "DESC"
        }, {
          "name" : "filter",
          "in" : "query",
          "description" : "Filter in JSON format",
          "example" : {
            "visibility" : "PUBLIC",
            "category" : "GENERAL"
          }
        }, {
          "name" : "q",
          "in" : "query",
          "description" : "Quick search query",
          "example" : "vocabulary"
        }, {
          "name" : "searchFields",
          "in" : "query",
          "description" : "Fields to search in (comma-separated)",
          "example" : "name,description"
        }, {
          "name" : "category.eq",
          "in" : "query",
          "description" : "Filter category equals",
          "example" : "GENERAL"
        }, {
          "name" : "visibility.eq",
          "in" : "query",
          "description" : "Filter visibility equals",
          "example" : "PUBLIC"
        }, {
          "name" : "createdAt.gt",
          "in" : "query",
          "description" : "Filter created after date",
          "example" : "2023-01-01"
        }, {
          "name" : "likeCount.gte",
          "in" : "query",
          "description" : "Filter minimum likes",
          "example" : 10
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseVocabularySetDto"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Create a new vocabulary set",
        "description" : "Creates a new vocabulary set with the provided information",
        "operationId" : "createVocabularySet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseOptionalVocabularySetDto"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/v1/vocabulary-sets/recent" : {
      "get" : {
        "tags" : [ "Vocabulary Sets" ],
        "summary" : "Get recently used vocabulary sets",
        "description" : "Retrieves all recently used vocabulary sets",
        "operationId" : "getRecentlyUsedVocabularySets",
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ApiResponseObject" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          },
          "meta" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Link"
            }
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "Link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string"
          },
          "rel" : {
            "type" : "string"
          },
          "method" : {
            "type" : "string"
          }
        }
      },
      "UpdateRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PRIVATE", "UNLISTED", "PUBLIC" ]
          },
          "difficultyLevel" : {
            "type" : "string",
            "enum" : [ "BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT" ]
          },
          "category" : {
            "type" : "string",
            "enum" : [ "GENERAL", "ACADEMIC", "BUSINESS", "TECHNOLOGY", "MEDICAL", "IELTS", "TOEIC", "TOEFL", "TRAVEL", "FOOD_AND_DRINK", "SCIENCE", "ARTS", "SPORTS", "ENTERTAINMENT", "SOCIAL_MEDIA", "CULTURE", "HISTORY", "GEOGRAPHY", "LANGUAGE", "PSYCHOLOGY", "PHILOSOPHY", "RELIGION", "POLITICS", "TECHNICAL", "FINANCE", "ENVIRONMENT", "EDUCATION", "DAILY_CONVERSATION", "LITERATURE", "MUSIC", "OTHER" ]
          }
        }
      },
      "ApiResponseVocabularySetDto" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/VocabularySetDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          },
          "meta" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Link"
            }
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "VocabularySetDto" : {
        "required" : [ "name", "visibility" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Unique identifier of the vocabulary set",
            "format" : "uuid"
          },
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string",
            "description" : "Name of the vocabulary set"
          },
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string",
            "description" : "Description of the vocabulary set"
          },
          "isActive" : {
            "type" : "boolean",
            "writeOnly" : true
          },
          "createdBy" : {
            "type" : "string",
            "description" : "ID of the user who created the vocabulary set"
          },
          "viewCount" : {
            "type" : "integer",
            "description" : "Number of views",
            "format" : "int64"
          },
          "likeCount" : {
            "type" : "integer",
            "description" : "Number of likes",
            "format" : "int64"
          },
          "shareCount" : {
            "type" : "integer",
            "description" : "Number of shares",
            "format" : "int64"
          },
          "visibility" : {
            "type" : "string",
            "description" : "Visibility level of the vocabulary set",
            "enum" : [ "PRIVATE", "UNLISTED", "PUBLIC" ]
          },
          "difficultyLevel" : {
            "type" : "string",
            "description" : "Difficulty level of the vocabulary set",
            "enum" : [ "BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT" ]
          },
          "category" : {
            "type" : "string",
            "description" : "Category of the vocabulary set",
            "enum" : [ "GENERAL", "ACADEMIC", "BUSINESS", "TECHNOLOGY", "MEDICAL", "IELTS", "TOEIC", "TOEFL", "TRAVEL", "FOOD_AND_DRINK", "SCIENCE", "ARTS", "SPORTS", "ENTERTAINMENT", "SOCIAL_MEDIA", "CULTURE", "HISTORY", "GEOGRAPHY", "LANGUAGE", "PSYCHOLOGY", "PHILOSOPHY", "RELIGION", "POLITICS", "TECHNICAL", "FINANCE", "ENVIRONMENT", "EDUCATION", "DAILY_CONVERSATION", "LITERATURE", "MUSIC", "OTHER" ]
          },
          "isSystem" : {
            "type" : "boolean",
            "writeOnly" : true
          },
          "vocabularyCount" : {
            "type" : "integer",
            "description" : "Number of vocabularies in the set",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "description" : "Creation timestamp",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "description" : "Last update timestamp",
            "format" : "date-time"
          },
          "isOwner" : {
            "type" : "boolean",
            "writeOnly" : true
          },
          "isLiked" : {
            "type" : "boolean",
            "writeOnly" : true
          },
          "active" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "boolean"
          },
          "system" : {
            "type" : "boolean"
          },
          "liked" : {
            "type" : "boolean"
          }
        },
        "description" : "Vocabulary Set Data Transfer Object"
      },
...
      "CreateRequest" : {
        "required" : [ "name", "visibility" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PRIVATE", "UNLISTED", "PUBLIC" ]
          },
          "difficultyLevel" : {
            "type" : "string",
            "enum" : [ "BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT" ]
          },
          "category" : {
            "type" : "string",
            "enum" : [ "GENERAL", "ACADEMIC", "BUSINESS", "TECHNOLOGY", "MEDICAL", "IELTS", "TOEIC", "TOEFL", "TRAVEL", "FOOD_AND_DRINK", "SCIENCE", "ARTS", "SPORTS", "ENTERTAINMENT", "SOCIAL_MEDIA", "CULTURE", "HISTORY", "GEOGRAPHY", "LANGUAGE", "PSYCHOLOGY", "PHILOSOPHY", "RELIGION", "POLITICS", "TECHNICAL", "FINANCE", "ENVIRONMENT", "EDUCATION", "DAILY_CONVERSATION", "LITERATURE", "MUSIC", "OTHER" ]
          },
          "vocabularyIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }
      },
      "ApiResponseOptionalVocabularySetDto" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/VocabularySetDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          },
          "meta" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Link"
            }
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponsePagedResponseVocabularySetDto" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseVocabularySetDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          },
          "meta" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Link"
            }
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseVoid" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          },
          "meta" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Link"
            }
          },
          "success" : {
            "type" : "boolean"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}